Extract desktop package for running desktop CLI
Previously, there was an overabundance of types and helper functions in
`desktop_launch.go` and it was becoming hard to determine which were
helpers for running and parsing `flight desktop` and which were helpers
for processing and rendering the web pages/form data.

The types and functions related to running `flight desktop` have been
extracted to a `desktop` package and the minimal set of names exported.
